Api to UI

+++ app/api/posts/[id]/route.ts
@@ -1,5 +1,7 @@
 import prisma from "@/lib/prismadb";
+import { getServerSession } from "next-auth/next";
 import { NextResponse } from "next/server";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 export async function GET(
     req: Request,
@@ -19,6 +21,14 @@ export async function PUT(
     req: Request,
     { params }: { params: { id: string } }
 ) {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json(
+            { error: "Not authenticated" },
+            { status: 401 }
+        );
+    }
+
     const { title, content, links, selectedCategory, imageUrl, publicId } =
         await req.json();
     const id = params.id;
@@ -45,6 +55,14 @@ export async function DELETE(
     req: Request,
     { params }: { params: { id: string } }
 ) {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json(
+            { error: "Not authenticated" },
+            { status: 401 }
+        );
+    }
+
     const id = params.id;
     try {
         const post = await prisma.post.delete({ where: { id } });
